backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
pymupdf==1.23.4
langdetect==1.0.9
pydantic==2.4.2
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
sentence-transformers==2.2.2
chromadb==0.4.15
loguru==0.7.2
python-docx==1.1.0
pywin32==306
rank_bm25==0.2.2
g4f==0.2.7.2
